refactor(cli): flatten connector command module and shared test helpers
Move the Connector command slice and its shared test helpers to neutral command paths, update every current caller, and preserve all compatibility behavior.
