+++ turbo/apps/cli/src/commands/connector/custom/__tests__/create.test.ts
@@ -6,7 +6,7 @@ import chalk from "chalk";
 import { http, HttpResponse } from "msw";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { server } from "../../../../../mocks/server";
+import { server } from "../../../../mocks/server";
 import { customConnector } from "../../../__tests__/helpers/custom-connectors";
 import { createCustomConnectorCommand } from "../create";
 import { customConnectorCommand } from "../index";
@@ -114,7 +114,7 @@ describe("okou connector custom create", () => {
 
   beforeEach(() => {
     chalk.level = 0;
-    tempDir = mkdtempSync(join(tmpdir(), "zero-custom-connector-create-"));
+    tempDir = mkdtempSync(join(tmpdir(), "custom-connector-create-"));
     vi.stubEnv("VM0_API_BACKEND_URL", "http://localhost:3000");
     vi.stubEnv("OKOU_TOKEN", buildZeroToken(["connector:write"]));
     vi.stubEnv("OKOU_AGENT_ID", AGENT_ID);
