+++ turbo/apps/cli/src/commands/connector/custom/__tests__/index.test.ts
@@ -3,7 +3,7 @@ import { http, HttpResponse } from "msw";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import type { CustomConnectorMcpResponse } from "@okouai/api-contracts/contracts/zero-custom-connectors";
 
-import { server } from "../../../../../mocks/server";
+import { server } from "../../../../mocks/server";
 import {
   customConnector,
   stubAgentCustomConnectors,
