+++ turbo/apps/cli/src/commands/generate/__tests__/lister.test.ts
@@ -9,7 +9,7 @@ import {
   manualAuthMethod,
   stubConnectorCatalog,
   stubConnectorCatalogStatus,
-} from "../../zero/__tests__/helpers/connector-catalog";
+} from "../../__tests__/helpers/connector-catalog";
 
 const AGENT_ID = "550e8400-e29b-41d4-a716-446655440000";
